refactor: Implementados controladores y servicios para Users

+++ src/main/java/es/daw2/fct_fct/repositorio/RepositorioUser.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import es.daw2.fct_fct.modelo.Users;
 
-public interface RepositorioUser extends CrudRepository<Users, Integer> {
+public interface RepositorioUser extends CrudRepository<Users, Long> {
     
 }
